fix profil typo
